+++ SFileHub-Back/src/test/java/com/keyon/sfilehub/T00001CreateUserTest.java
@@ -1,6 +1,5 @@
 package com.keyon.sfilehub;
 
-import com.alibaba.fastjson2.JSONObject;
 import com.keyon.sfilehub.dao.RoleDao;
 import com.keyon.sfilehub.dao.UserDao;
 import com.keyon.sfilehub.entity.Role;
@@ -9,27 +8,14 @@ import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.ResponseCookie;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.test.web.reactive.server.EntityExchangeResult;
-import org.springframework.test.web.reactive.server.WebTestClient;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.reactive.function.BodyInserters;
 
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureWebTestClient
+@SpringBootTest
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class HelloControllerTest {
-    @LocalServerPort
-    private int port;
-    @Autowired
-    private WebTestClient webTestClient;
+public class T00001CreateUserTest {
 
     @Autowired
     private UserDao userDao;
@@ -40,11 +26,10 @@ public class HelloControllerTest {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
-    private static MultiValueMap<String, ResponseCookie> cookies;
-
     @Test
     @Order(1)
     public void addRole() {
+        System.out.println("11111");
         if (roleDao.findByTitle("ROLE_ADMIN") != null) return;
         Role role = Role.builder().title("ROLE_ADMIN").label("管理员").intro("最高权限者").enable(true).build();
         roleDao.save(role);
@@ -91,50 +76,5 @@ public class HelloControllerTest {
         roles.add(userRole);
         user.setRoles(roles);
         userDao.save(user);
-    }
-
-    @Test
-    @Order(3)
-    @DisplayName("login admin")
-    void login() {
-        Map<String, String> bodyMap = new HashMap<>();
-        bodyMap.put("username", "admin");
-        bodyMap.put("password", "123456");
-        String url = "http://localhost:" + port + "/login";
-        EntityExchangeResult<String> resp = webTestClient
-                .post()
-                .uri(url)
-                .body(BodyInserters.fromValue(bodyMap))
-                .exchange()
-                .expectStatus().isOk()
-                .expectBody(String.class).returnResult();
-        String result = resp.getResponseBody();
-        System.out.println("login res: " + result);
-        JSONObject jsonResult = JSONObject.parseObject(result);
-        Assertions.assertNotNull(jsonResult);
-        Assertions.assertNotNull(jsonResult.get("code"));
-        Assertions.assertEquals((int) jsonResult.get("code"), 0);
-        cookies = resp.getResponseCookies();
-        System.out.println("cookies key: " + cookies.keySet());
-        Assertions.assertEquals(cookies.keySet().toString(), "[JSESSIONID]");
-    }
-
-
-    @Test
-    @Order(4)
-    @DisplayName("a hello test")
-    void greetingTest() {
-        System.out.println("cookies key2: " + cookies.keySet());
-        String url = "http://localhost:" + port + "/hello/get";
-        EntityExchangeResult<String> resp = webTestClient
-                .get()
-                .uri(url)
-                .cookie("JSESSIONID", cookies.get("JSESSIONID").get(0).getValue())
-                .exchange()
-                .expectStatus().isOk()
-                .expectBody(String.class)
-                .returnResult();
-        String result = resp.getResponseBody();
-        Assertions.assertEquals(result, "Hello World!");
     }
 }
